phantomjs os

+++ DataMinner/common.py
@@ -28,13 +28,30 @@ class CoinData:
 
 def phantom_get(url):
     url_file = open('d:/url.txt', 'w')
+    if os.path.exists('d:/flag'):
+        os.remove('d:/flag')
     url_file.write(url)
     url_file.flush()
     url_file.close()
-    shutil.copy("phantomGet.js", "d:/phantomGet.js")
-    time.sleep(1)
+    if not os.path.exists("d:/phantomGet.js"):
+        shutil.copy("phantomGet.js", "d:/phantomGet.js")
+        time.sleep(1)
     os.system("phantomjs d:/phantomGet.js")
+    ok = False
+    for i in range(300):
+        if os.path.exists('d:/flag'):
+            os.remove('d:/flag')
+            ok = True
+            break
+        else:
+            time.sleep(1)
+    if ok:
+        file = open('d:/output.html', 'r', encoding='utf-8')
+        return file.read()
+    else:
+        return None
 
 
 if __name__ == "__main__":
-    phantom_get("https://www.baidu.com")
+    result = phantom_get("https://www.huobi.com")
+    print(result)
